content-marketing-agent/requirements.txt
streamlit
openai
pydantic
fastapi
uvicorn
dspy-ai
deep-translator
langdetect
apify-client
httpx
requests
beautifulsoup4
jinja2
aiofiles
python-dotenv